Add handy LimitedHashMap class for caches
